move types around for shared flows
