+++ src/features/DeFiZap/helpers.ts
@@ -1,9 +1,7 @@
-import { IHexStrTransaction, ITxObject } from '@types';
+import { IHexStrTransaction, ITxObject, ISimpleTxForm } from '@types';
 import { inputValueToHex, inputGasPriceToHex, inputNonceToHex } from '@services/EthService';
 import { DEFAULT_NETWORK_CHAINID } from '@config';
 import { COLORS } from '@theme';
-
-import { ISimpleTxForm } from './types';
 
 export const createSimpleTxObject = (formData: ISimpleTxForm): IHexStrTransaction | ITxObject => {
   return {
