+++ src/features/PurchaseMembership/types.ts
@@ -1,7 +1,5 @@
-import { Asset, Network, StoreAccount } from '@types';
+import { StoreAccount, ISimpleTxFormFull } from '@types';
 import { IMembershipConfig } from './config';
-
-export type TStepAction = (payload: any, after: () => void) => void;
 
 export interface IMembershipStepComponentProps {
   completeButtonText: string;
@@ -12,20 +10,6 @@ export interface IMembershipStepComponentProps {
 export interface MembershipPurchaseState {
   membershipSelected: undefined | IMembershipConfig; // @todo: Make enum
   account: StoreAccount;
-}
-
-export interface ISimpleTxForm {
-  address: string; // simple eth address
-  amount: string; // in ether - ex: 1
-  gasLimit: string | number; // number - ex: 1,500,000
-  gasPrice: string; // gwei
-  nonce: string; // number - ex: 55
-  account: StoreAccount;
-}
-
-export interface ISimpleTxFormFull extends ISimpleTxForm {
-  asset: Asset;
-  network: Network;
 }
 
 export interface MembershipSimpleTxFormFull extends ISimpleTxFormFull {
